# GAIA_GPTs_Adaptados/37_OPENAPI_GPT_RADAR_CNPJ.yaml
openapi: 3.1.0
info:
  title: GAIA Radar - CNPJ e Funil Comercial
  version: 2.0.0
  description: Consulta dados cadastrais públicos e registra empresas no Funil GAIA.
servers:
  - url: https://gaia-radar-api.vercel.app
paths:
  /api/empresa/consultar:
    post:
      operationId: consultarEmpresaPorCnpj
      summary: Consulta uma empresa brasileira pelo CNPJ.
      description: Use quando o usuário fornecer um CNPJ e pedir pesquisa, qualificação ou dossiê comercial.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                cnpj:
                  type: string
                  description: CNPJ com ou sem pontuação.
              required:
                - cnpj
      responses:
        "200":
          description: Dados cadastrais públicos normalizados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmpresaPublica"
        "400":
          description: CNPJ inválido.
        "401":
          description: Chave da Ação ausente ou inválida.
        "404":
          description: CNPJ não encontrado.
        "429":
          description: Limite temporário da fonte pública atingido.
  /api/funil:
    get:
      operationId: consultarFunilGaia
      summary: Consulta empresas e próximas ações salvas no Funil GAIA.
      description: Use quando o usuário pedir para ver o funil, localizar uma empresa, listar pendências ou consultar uma etapa comercial.
      security:
        - bearerAuth: []
      parameters:
        - name: busca
          in: query
          required: false
          description: Nome, CNPJ ou termo relacionado à empresa.
          schema:
            type: string
        - name: etapa
          in: query
          required: false
          description: Etapa comercial a consultar.
          schema:
            type: string
            enum: [prospectivo, contato, visita, analise, proposta, cliente]
      responses:
        "200":
          description: Empresas encontradas no Funil GAIA.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunilLista"
        "401":
          description: Chave da Ação ausente ou inválida.
        "503":
          description: Armazenamento em nuvem ainda não configurado.
    post:
      operationId: salvarEmpresaNoFunil
      summary: Salva ou atualiza uma empresa no Funil GAIA.
      description: Use somente quando o usuário pedir para salvar, registrar, atualizar etapa ou definir a próxima ação de uma empresa. Nunca salve dados financeiros sensíveis.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                id:
                  type: string
                  description: Identificador existente, quando estiver atualizando um registro consultado.
                empresa:
                  type: string
                cnpj:
                  type: string
                etapa:
                  type: string
                  enum: [prospectivo, contato, visita, analise, proposta, cliente]
                proximaAcao:
                  type: string
                dataProximaAcao:
                  type: string
                  description: Data no formato AAAA-MM-DD.
                observacoes:
                  type: string
                dossie:
                  type: string
                  description: Resumo do dossiê com apenas dados públicos.
              required:
                - empresa
      responses:
        "200":
          description: Empresa salva ou atualizada.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  entry:
                    $ref: "#/components/schemas/FunilEntry"
                  created:
                    type: boolean
        "400":
          description: Dados inválidos.
        "401":
          description: Chave da Ação ausente ou inválida.
        "503":
          description: Armazenamento em nuvem ainda não configurado.
    delete:
      operationId: excluirEmpresaDoFunil
      summary: Exclui uma empresa do Funil GAIA.
      description: Use apenas após o usuário pedir explicitamente a exclusão de uma empresa já consultada no funil.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Empresa excluída.
        "401":
          description: Chave da Ação ausente ou inválida.
        "404":
          description: Empresa não encontrada.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    FunilEntry:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        empresa:
          type: string
        cnpj:
          type: string
        etapa:
          type: string
          enum: [prospectivo, contato, visita, analise, proposta, cliente]
        proximaAcao:
          type: string
        dataProximaAcao:
          type: string
        observacoes:
          type: string
        dossie:
          type: string
        criadoEm:
          type: string
        atualizadoEm:
          type: string
      required: [id, empresa, cnpj, etapa, proximaAcao, dataProximaAcao, observacoes, dossie, criadoEm, atualizadoEm]
    FunilLista:
      type: object
      additionalProperties: false
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/FunilEntry"
        total:
          type: integer
        cloud:
          type: boolean
      required: [entries, total, cloud]
    EmpresaPublica:
      type: object
      additionalProperties: false
      properties:
        cnpj:
          type: string
        razaoSocial:
          type: string
        nomeFantasia:
          type: [string, "null"]
        situacaoCadastral:
          type: string
        dataSituacaoCadastral:
          type: [string, "null"]
        dataInicioAtividade:
          type: [string, "null"]
        porte:
          type: [string, "null"]
        naturezaJuridica:
          type: [string, "null"]
        capitalSocial:
          type: [number, "null"]
        atividadePrincipal:
          type: [string, "null"]
        atividadesSecundarias:
          type: array
          items:
            type: string
        endereco:
          type: object
          additionalProperties: false
          properties:
            logradouro:
              type: [string, "null"]
            bairro:
              type: [string, "null"]
            cidade:
              type: [string, "null"]
            estado:
              type: [string, "null"]
            cep:
              type: [string, "null"]
        canaisOficiais:
          type: object
          additionalProperties: false
          properties:
            telefones:
              type: array
              items:
                type: string
            email:
              type: [string, "null"]
        socios:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              nome:
                type: string
              tipo:
                type: [string, "null"]
              qualificacao:
                type: [string, "null"]
        fonte:
          type: object
          additionalProperties: false
          properties:
            nome:
              type: string
            url:
              type: string
            consultadoEm:
              type: string
            atualizadoEm:
              type: [string, "null"]
        aviso:
          type: string
      required:
        - cnpj
        - razaoSocial
        - situacaoCadastral
        - atividadesSecundarias
        - endereco
        - canaisOficiais
        - socios
        - fonte
        - aviso